requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
et_xmlfile==2.0.0
ftputil==5.1.0
future==1.0.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.6.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
paramiko==3.5.0
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
textfsm==1.1.3
tzdata==2025.1
xlrd==2.0.1
